+++ client/package.json
@@ -24,6 +24,7 @@
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.2.0",
+    "redux-thunk": "^2.3.0",
     "simple-peer": "^9.7.2",
     "socket.io-client": "^2.3.0",
     "sweetalert2": "^9.17.0"
