+++ client/src/pages/PomodoroPage.jsx
@@ -1,9 +1,7 @@
 import React from 'react';
-import PomodoroVideo from '../components/PomodoroVideo';
 import PomodoroTimer from '../components/PomodoroTimer';
 import PomodoroMessage from '../components/PomodoroMessage';
 import PomodoroUserInfo from '../components/PomodoroUserInfo';
-import MatchingSystem from "../components/MatchingSystem";
 
 const styles = {
   videoBox: {
@@ -26,12 +24,9 @@ export default function PomodoroPage() {
       <PomodoroTimer />
       <div css={styles.videoBox}>
         <div css={styles.userBox}>
-          {/*<PomodoroVideo />*/}
-          <MatchingSystem />
           <PomodoroUserInfo />
         </div>
         <div css={styles.userBox}>
-          <PomodoroVideo />
           <PomodoroUserInfo />
         </div>
       </div>
