+++ client/src/store.js
@@ -1,9 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit';
 
+import thunk from 'redux-thunk';
 import reducer from './slice';
 
 const store = configureStore({
   reducer,
+  middleware: [thunk],
 });
 
 export default store;
